Refactor models -> queue & user modules
